+++ src/scroll.ts
@@ -1,14 +1,16 @@
 import { AnimationManager } from "./animation-manager"
 import { ScrollElement, IHorizontal, toDirection, Point } from "./element"
 import { EasingFunction, defaultEasingFunction } from "./default-settings"
+import { Misc } from "./misc"
 
 function almost0(value: number): boolean {
   return value < 1 && value > -1
 }
 
 class Scroll {
-  private element: ScrollElement
+  element: ScrollElement
   private animationManager: IHorizontal<AnimationManager>
+  misc: Misc
   constructor(
     element: Element | Window = window,
     public onScroll: ((external?: boolean) => void) | null = null,
@@ -22,18 +24,42 @@ class Scroll {
       this.onScroll && this.onScroll(!(almostX && almostY))
     })
 
+    this.misc = new Misc(this.element)
+
     this.animationManager = {
       x: new AnimationManager(this.element.position.x),
       y: new AnimationManager(this.element.position.y),
     }
 
-    this.scroll()
+    const scroll = () => {
+      const shouldBe = new Point(this.animationManager.x.shouldBe, this.animationManager.y.shouldBe)
+      this.animationManager.x.updateShouldBe()
+      this.animationManager.y.updateShouldBe()
+      if (
+        shouldBe.x !== this.animationManager.x.shouldBe ||
+        shouldBe.y !== this.animationManager.y.shouldBe
+      ) {
+        this.element.scrollTo(
+          new Point(this.animationManager.x.shouldBe, this.animationManager.y.shouldBe),
+        )
+      }
+      this.animationManager.x.shouldBe = Math.max(
+        0,
+        Math.min(this.animationManager.x.shouldBe, this.element.scrollSize.x - this.element.size.x),
+      )
+      this.animationManager.y.shouldBe = Math.max(
+        0,
+        Math.min(this.animationManager.y.shouldBe, this.element.scrollSize.y - this.element.size.y),
+      )
+      window.requestAnimationFrame(scroll)
+    }
+    window.requestAnimationFrame(scroll)
   }
-  public stopAllAnimations() {
+  stopAllAnimations() {
     this.animationManager.x.stopAllAnimations()
     this.animationManager.y.stopAllAnimations()
   }
-  public scrollBy(
+  scrollBy(
     distToScroll: number,
     duration: number,
     horizontal: boolean = false,
@@ -45,48 +71,6 @@ class Scroll {
       duration,
       easing,
     })
-  }
-  private scroll = () => {
-    const shouldBeX = this.animationManager.x.shouldBe
-    const shouldBeY = this.animationManager.y.shouldBe
-    this.animationManager.x.updateShouldBe()
-    this.animationManager.y.updateShouldBe()
-    if (
-      shouldBeX !== this.animationManager.x.shouldBe ||
-      shouldBeY !== this.animationManager.y.shouldBe
-    ) {
-      this.element.scrollTo(
-        new Point(this.animationManager.x.shouldBe, this.animationManager.y.shouldBe),
-      )
-    }
-    this.animationManager.x.shouldBe = Math.max(
-      0,
-      Math.min(this.animationManager.x.shouldBe, this.element.scrollSize.x - this.element.size.x),
-    )
-    this.animationManager.y.shouldBe = Math.max(
-      0,
-      Math.min(this.animationManager.y.shouldBe, this.element.scrollSize.y - this.element.size.y),
-    )
-    window.requestAnimationFrame(this.scroll)
-  }
-  // private getDist(scrollType: ScrollType, value: number, horizontal: boolean): number {
-  //   const direction = toDirection(horizontal)
-  //   switch (scrollType) {
-  //     case "percent":
-  //       return ((this.element.scrollSize[direction] - this.element.size[direction]) * value) / 100
-  //     case "screen":
-  //       return this.element.size[direction] * value
-  //     default:
-  //       return value
-  //   }
-  // }
-  public getDistToElement(element: Element, value: number, horizontal: boolean): number {
-    const ratio = value / 100
-    const direction = toDirection(horizontal)
-    const elementWrapper = new ScrollElement(element)
-    const screenOffset = (this.element.size[direction] - elementWrapper.size[direction]) * ratio
-    const elementPosition = elementWrapper.offset[direction] - this.element.offset[direction]
-    return elementPosition - screenOffset
   }
 }
